autorest level rename of client
