+++ sdk/farmbeats/azure-farmbeats/azure/farmbeats/models/_farm_beats_client_enums.py
